stoped in strlcat

+++ ft_bzero.c
@@ -10,10 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*str = s;
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
