+++ ft_strjoin.c
@@ -10,12 +10,14 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	 *ft_strjoin(char const *s1, char const *s2)
 {
 	char	*str;
 	int	i;
 	int	j;
-	int	len;
+	size_t	len;
 
 	i = 0;
 	j = 0;
